signing: merge sign_transaction_essence()
Avoid duplication by reusing the loop going through transaction
inputs and outputs. To achieve this without removing it from a
`Signer`, `sign_transaction_essence()` is provided as a default
implementation of `Signer`. `signature_unlock()` is added to
`Signer` for performing the actual signs.
